revert: reverting part of f0453c
The currently used v8z does not have the api for
SetAbortOnUncaughtExceptionCallback. Below is a description
of the commit that added the SetAbortOnUncaughtExceptionCallback.

f0453c includes the below port from v0.10 to v0.12

caeb677
Do not abort the process if an error is thrown from within a domain,
an error handler is setup for the domain and
--abort-on-uncaught-exception was passed on the command line.

However, if an error is thrown from within the top-level domain's
error handler and --abort-on-uncaught-exception was passed on the
command line, make the process abort.