Get same ip for connect and listen on localhost.
Connect and listen do not follow the same code path when they require
dns lookup on a host and port. On z/OS this results in different ips
from getaddrinfo when host is 'localhost' causing ping pong test to fail
on local host. Change listen to use dns options like connect so
getaddrinfo is called with the same hints and returns the same addrinfo.