[z/OS] Skip character conversion for ASCII streams
Node on z/OS operates under the assumption that most programs
will generate EBCDIC output on stdout/stderr given that its
an EBCDIC based system. However there are programs which are
compiled with -qascii which produce ASCII on stdout/stderr. For
these programs we will skip EBCDIC->ASCII conversion prior to
consumption in Node. In the absence of mechanism to detect -qascii
programs we will operate with a static list for now which include
git and python.