Fix x509_cmp_time digits check for z/OS
x509_cmp_time uses the isDigit function from ctypes to validate
the time.  On zOS, this does an EBCDIC test, while the strings
when compiled in Node are treated as ASCII.  Convert the conditional
to an explicit test for the ASCII digit range.

Fix encoding format for generalizedtime_length as well.

Signed-off-by: Joran Siu <joransiu@ca.ibm.com>